Consolidate module location logic into module-paths
